Optimize get_label to run in O(n)
When generating labels from the train.csv, the script is very slow for large categories. The culprit seems to be the result of a very slow dataframe access pattern. By caching the properly filtered list, we can reduce the runtime from O(n^2) to O(n). This massively speeds up label generation.